Move functions of user accepting from enter-room/rtdb to accept-user/rtdb
